+++ mcpath/_dummy.py
@@ -1,6 +1,10 @@
-__all__ = ["java", "bedrock", "preview", "education"]
+__all__ = ["java", "bedrock", "preview", "education", "get_runtime"]
 
 java = None
 bedrock = None
 preview = None
 education = None
+
+
+def get_runtime(version: str) -> None:
+    return None