+++ mcpath/_windows.py
@@ -1,51 +1,93 @@
 """Windows (win32) paths"""
 
 import os
+from .facade import JEPath, BEPath, PREPath, EDUPath
+from .utils import _get_latest_profile, _version_to_component
 
-__all__ = ["java", "bedrock", "preview", "education"]
-
-from mcpath import JEPath, BEPath, PREPath, EDUPath
+__all__ = ["java", "bedrock", "preview", "education", "get_runtime"]
 
 
-class WindowsJEPath(JEPath):
+class WinJEPath(JEPath):
     platform = "win32"
 
-    @property
-    def root(self) -> str:
-        return os.path.expandvars("%appdata%\\.minecraft")
+    def _get_launcher(self) -> str:
+        path = os.path.join(
+            "C:\\" + "XboxGames", "Minecraft Launcher", "Content", "Minecraft.exe"
+        )
+        return path
+
+    def _get_executable(self) -> str:
+        return os.path.expandvars("%APPDATA%\\.minecraft\\versions")
+
+    def _get_game(self) -> str:
+        fp = os.path.expandvars("%APPDATA%\\.minecraft\\launcher_profiles.json")
+        path = _get_latest_profile(fp)
+        if path:
+            return path
+        # fallback
+        return os.path.expandvars("%APPDATA%\\.minecraft")
 
 
-class WindowsBEPath(BEPath):
+class WinBEPath(BEPath):
     platform = "win32"
 
-    @property
-    def root(self) -> str:
+    def _get_game(self) -> str:
         return os.path.expandvars(
-            "%localappdata%\\Packages\\Microsoft.MinecraftUWP_8wekyb3d8bbwe\\LocalState\\games\\com.mojang"
+            "%LOCALAPPDATA%\\Packages\\Microsoft.MinecraftUWP_8wekyb3d8bbwe\\LocalState\\games\\com.mojang"
         )
 
 
-class WindowsPREPath(PREPath):
+class WinPREPath(PREPath):
     platform = "win32"
 
-    @property
-    def root(self) -> str:
+    def _get_game(self) -> str:
         return os.path.expandvars(
-            "%localappdata%\\Packages\\Microsoft.MinecraftWindowsBeta_8wekyb3d8bbwe\\LocalState\\games\\com.mojang"
+            "%LOCALAPPDATA%\\Packages\\Microsoft.MinecraftWindowsBeta_8wekyb3d8bbwe\\LocalState\\games\\com.mojang"
         )
 
 
-class WindowsEDUPath(EDUPath):
+class WinEDUPath(EDUPath):
     platform = "win32"
 
-    @property
-    def root(self) -> str:
+    def _get_game(self) -> str:
         return os.path.expandvars(
-            "%localappdata%\\Packages\\Microsoft.MinecraftEducationEdition_8wekyb3d8bbwe\\LocalState\\games\\com.mojang"
+            "%LOCALAPPDATA%\\Packages\\Microsoft.MinecraftEducationEdition_8wekyb3d8bbwe\\LocalState\\games\\com.mojang"
         )
 
 
-java = WindowsJEPath()
-bedrock = WindowsBEPath()
-preview = WindowsPREPath()
-education = WindowsEDUPath()
+def get_runtime(version: str) -> str:
+    """
+    Get the full path to the java runtime environment.
+
+    :param version: The version of Minecraft.
+    :type version: str
+    :return: The java runtime path.
+    :rtype: str
+    """
+    component, major_version = _version_to_component(version)
+    if component is None:
+        return "java"
+    path = os.path.expandvars(
+        os.path.join(
+            "%localappdata%",
+            "Packages",
+            "Microsoft.4297127D64EC6_8wekyb3d8bbwe",
+            "LocalCache",
+            "Local",
+            "runtime",
+            component,
+            "windows-x64",
+            component,
+            "bin",
+            "java.exe",
+        )
+    )
+    if os.path.isfile(path):
+        return path
+    return "java"
+
+
+java = WinJEPath()
+bedrock = WinBEPath()
+preview = WinPREPath()
+education = WinEDUPath()