+++ tests/test_bedrock.py
@@ -1,0 +1,5 @@
+import mcpath
+
+
+def test_bedrock():
+    print(mcpath.bedrock)