+++ tests/test_java.py
@@ -1,0 +1,5 @@
+import mcpath
+
+
+def test_java():
+    print(mcpath.java)